Factor out functional tests
